collectors/book_collector/requirements.txt
APScheduler
binance-connector
httpx
redis